FastAPI/requirements.txt
cryptography
fastapi
pyodbc
pydantic
uvicorn
python-dotenv